# Optimize rendering of ThreadStateBar