--- a/src/ClientData/include/ClientData/CaptureData.h
+++ b/src/ClientData/include/ClientData/CaptureData.h
@@ -127,7 +127,7 @@ class CaptureData {
       uint32_t thread_id, uint64_t min_timestamp, uint64_t max_timestamp,
       const std::function<void(const ThreadStateSliceInfo&)>& action) const;
 
-  // Similar to the previous one, but does not iterate for more than one slice per pixel.
+  // Similar to the previous one, but does not iterate over more than one slice per pixel.
   void ForEachThreadStateSliceIntersectingTimeRangeDiscretized(
       uint32_t thread_id, uint64_t min_timestamp, uint64_t max_timestamp, uint32_t resolution,
       const std::function<void(const ThreadStateSliceInfo&)>& action) const;